Minor Fixes

+++ QuickStop.Infrastructure/Serializers/HotelSerializer.cs
@@ -26,48 +26,45 @@ namespace QuickStop.Infrastructure.Serializers
             List<HotelRoom> hotels = new List<HotelRoom>();
 
             string file = FilePath(fileName);
-            try
-            {
-                string[] data = File.ReadAllLines(file);
 
-                foreach (string hotelData in data)
-                {
-                    string[] hotelParams = parser.Split(hotelData);
-
-                    try
-                    {
-                        HotelRoom hotel = new HotelRoom
-                        {
-                            ID = Convert.ToInt32(hotelParams[0]),
-                            Name = hotelParams[1].Trim('\"'),
-                            Description = hotelParams[2].Trim('\"'),
-                            Price = Convert.ToDecimal(hotelParams[3]),
-                            Ratings = Convert.ToUInt16(hotelParams[4]),
-                            Location = (Location)Enum.Parse(typeof(Location), hotelParams[5]),
-                            MinGuestCount = Convert.ToInt32(hotelParams[6]),
-                            MaxGuestCount = Convert.ToInt32(hotelParams[7]),
-                            DateUntilAvailable = string.IsNullOrWhiteSpace(hotelParams[8]) ? null : (DateTime?)Convert.ToDateTime(hotelParams[8]),
-                            Icon = hotelParams[9],
-                            Room = hotelParams[10]
-                        };
-
-                        hotels.Add(hotel);
-                    }
-                    catch
-                    {
-                        
-                    }
-                }
-
-                return hotels;
-            }
-            catch
+            if(!File.Exists(file))
             {
                 string src = GetResourceFileContentAsString(fileName + extension);
                 File.WriteAllText(file, src);
-
-                throw new HotelDataNotFoundException();
             }
+
+            string[] data = File.ReadAllLines(file);
+
+            foreach (string hotelData in data)
+            {
+                string[] hotelParams = parser.Split(hotelData);
+
+                try
+                {
+                    HotelRoom hotel = new HotelRoom
+                    {
+                        ID = Convert.ToInt32(hotelParams[0]),
+                        Name = hotelParams[1].Trim('\"'),
+                        Description = hotelParams[2].Trim('\"'),
+                        Price = Convert.ToDecimal(hotelParams[3]),
+                        Ratings = Convert.ToUInt16(hotelParams[4]),
+                        Location = (Location)Enum.Parse(typeof(Location), hotelParams[5]),
+                        MinGuestCount = Convert.ToInt32(hotelParams[6]),
+                        MaxGuestCount = Convert.ToInt32(hotelParams[7]),
+                        DateUntilAvailable = string.IsNullOrWhiteSpace(hotelParams[8]) ? null : (DateTime?)Convert.ToDateTime(hotelParams[8]),
+                        Icon = hotelParams[9],
+                        Room = hotelParams[10]
+                    };
+
+                    hotels.Add(hotel);
+                }
+                catch
+                {
+
+                }
+            }
+
+            return hotels;
         }
 
         void IHotelSerializer.SerializeHotels(IEnumerable<HotelRoom> hotels)
